requirements.txt
PyGithub
playsound
GitPython
pyyaml
halo